Addressing comments vol1
