+++ sql/core/src/test/scala/org/apache/spark/sql/scripting/SqlScriptingInterpreterSuite.scala
@@ -17,65 +17,155 @@
 
 package org.apache.spark.sql.scripting
 
-import org.apache.spark.SparkFunSuite
-import org.apache.spark.sql.catalyst.expressions.{Alias, Literal}
-import org.apache.spark.sql.catalyst.plans.logical.{OneRowRelation, Project}
-import org.apache.spark.sql.catalyst.trees.Origin
+import org.apache.spark.sql.{AnalysisException, Dataset, QueryTest, Row}
+import org.apache.spark.sql.catalyst.QueryPlanningTracker
+import org.apache.spark.sql.test.SharedSparkSession
 
-class SqlScriptingInterpreterSuite extends SparkFunSuite {
+/**
+ * SQL Scripting interpreter tests.
+ * Output from the parser is provided to the interpreter.
+ * Output from the interpreter (iterator over executable statements) is then checked - statements
+ * are executed and output DataFrames are compared with expected outputs.
+ */
+class SqlScriptingInterpreterSuite extends QueryTest with SharedSparkSession {
   // Helpers
-  case class TestLeafStatement(testVal: String) extends LeafStatementExec {
-    override def reset(): Unit = ()
+  private def verifySqlScriptResult(
+      sqlText: String, expected: Seq[Seq[Row]]): Unit = {
+    val interpreter = SqlScriptingInterpreter()
+    val compoundBody = spark.sessionState.sqlParser.parseScript(sqlText)
+    val executionPlan = interpreter.buildExecutionPlan(compoundBody)
+    val result = executionPlan.flatMap {
+      case statement: SingleStatementExec =>
+        if (statement.consumed) {
+          None
+        } else {
+          Some(Dataset.ofRows(spark, statement.parsedPlan, new QueryPlanningTracker))
+        }
+      case _ => None
+    }.toArray
+
+    assert(result.length == expected.length)
+    result.zip(expected).foreach{ case (df, expectedAnswer) => checkAnswer(df, expectedAnswer)}
   }
-
-  case class TestNestedStatementIterator(statements: Seq[CompoundStatementExec])
-    extends CompoundNestedStatementIteratorExec(statements)
-
-  case class TestBody(statements: Seq[CompoundStatementExec])
-    extends CompoundBodyExec(statements)
-
-  case class TestSparkStatementWithPlan(testVal: String)
-    extends SingleStatementExec(
-      parsedPlan = Project(Seq(Alias(Literal(testVal), "condition")()), OneRowRelation()),
-      Origin(startIndex = Some(0), stopIndex = Some(testVal.length)),
-      isInternal = false)
 
   // Tests
-  test("test body - single statement") {
-    val iter = TestNestedStatementIterator(Seq(TestLeafStatement("one")))
-    val statements = iter.map {
-      case TestLeafStatement(v) => v
-      case _ => fail("Unexpected statement type")
-    }.toList
-
-    assert(statements === List("one"))
+  test("select 1") {
+    verifySqlScriptResult("SELECT 1;", Seq(Seq(Row(1))))
   }
 
-  test("test body - no nesting") {
-    val iter = TestNestedStatementIterator(
-      Seq(
-        TestLeafStatement("one"),
-        TestLeafStatement("two"),
-        TestLeafStatement("three")))
-    val statements = iter.map {
-      case TestLeafStatement(v) => v
-      case _ => fail("Unexpected statement type")
-    }.toList
-
-    assert(statements === Seq("one", "two", "three"))
+  test("multi statement - simple") {
+    withTable("t") {
+      val sqlScript =
+        """
+          |BEGIN
+          |CREATE TABLE t (a INT, b STRING, c DOUBLE) USING parquet;
+          |INSERT INTO t VALUES (1, 'a', 1.0);
+          |SELECT a, b FROM t WHERE a = 12;
+          |SELECT a FROM t;
+          |END
+          |""".stripMargin
+      val expected = Seq(
+        Seq.empty[Row], // create table
+        Seq.empty[Row], // insert
+        Seq.empty[Row], // select with filter
+        Seq(Row(1))
+      )
+      verifySqlScriptResult(sqlScript, expected)
+    }
   }
 
-  test("test body - nesting") {
-    val iter = TestNestedStatementIterator(
-      Seq(
-        TestNestedStatementIterator(Seq(TestLeafStatement("one"), TestLeafStatement("two"))),
-        TestLeafStatement("three"),
-        TestNestedStatementIterator(Seq(TestLeafStatement("four"), TestLeafStatement("five")))))
-    val statements = iter.map {
-      case TestLeafStatement(v) => v
-      case _ => fail("Unexpected statement type")
-    }.toList
+  test("multi statement - count") {
+    withTable("t") {
+      val sqlScript =
+        """
+          |BEGIN
+          |CREATE TABLE t (a INT, b STRING, c DOUBLE) USING parquet;
+          |INSERT INTO t VALUES (1, 'a', 1.0);
+          |INSERT INTO t VALUES (1, 'a', 1.0);
+          |SELECT
+          | CASE WHEN COUNT(*) > 10 THEN true
+          | ELSE false
+          | END AS MoreThanTen
+          |FROM t;
+          |END
+          |""".stripMargin
+      val expected = Seq(
+        Seq.empty[Row], // create table
+        Seq.empty[Row], // insert #1
+        Seq.empty[Row], // insert #2
+        Seq(Row(false))
+      )
+      verifySqlScriptResult(sqlScript, expected)
+    }
+  }
 
-    assert(statements === Seq("one", "two", "three", "four", "five"))
+  test("session vars - set and read") {
+    val sqlScript =
+      """
+        |BEGIN
+        |DECLARE var = 1;
+        |SET VAR var = var + 1;
+        |SELECT var;
+        |END
+        |""".stripMargin
+    val expected = Seq(
+      Seq.empty[Row], // declare var
+      Seq.empty[Row], // set var
+      Seq(Row(2)), // select
+      Seq.empty[Row] // drop var
+    )
+    verifySqlScriptResult(sqlScript, expected)
+  }
+
+  test("session vars - set and read scoped") {
+    val sqlScript =
+      """
+        |BEGIN
+        | BEGIN
+        |   DECLARE var = 1;
+        |   SELECT var;
+        | END;
+        | BEGIN
+        |   DECLARE var = 2;
+        |   SELECT var;
+        | END;
+        | BEGIN
+        |   DECLARE var = 3;
+        |   SET VAR var = var + 1;
+        |   SELECT var;
+        | END;
+        |END
+        |""".stripMargin
+    val expected = Seq(
+      Seq.empty[Row], // declare var
+      Seq(Row(1)), // select
+      Seq.empty[Row], // drop var
+      Seq.empty[Row], // declare var
+      Seq(Row(2)), // select
+      Seq.empty[Row], // drop var
+      Seq.empty[Row], // declare var
+      Seq.empty[Row], // set var
+      Seq(Row(4)), // select
+      Seq.empty[Row], // drop var
+    )
+    verifySqlScriptResult(sqlScript, expected)
+  }
+
+  test("session vars - var out of scope") {
+    val e = intercept[AnalysisException] {
+      val sqlScript =
+        """
+          |BEGIN
+          | BEGIN
+          |   DECLARE testVarName = 1;
+          |   SELECT testVarName;
+          | END;
+          | SELECT testVarName;
+          |END
+          |""".stripMargin
+      verifySqlScriptResult(sqlScript, Seq.empty)
+    }
+    assert(e.getErrorClass === "UNRESOLVED_COLUMN.WITHOUT_SUGGESTION")
+    assert{e.getMessage.contains("testVarName")}
   }
 }
